+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 //
+- added reflection invoke with return value
+- made logging generic
 
 ## [1.0.0] - 2024-01-22